+++ classes/api.php
@@ -393,6 +393,11 @@ class api {
      */
     public static function delete_all_data_for_course(int $courseid) {
         global $DB;
+        if ($context = \context_course::instance($courseid, IGNORE_MISSING)) {
+            // The context does not exist when this is called after the course was deleted,
+            // in this case the files in the course context were already deleted by core.
+            get_file_storage()->delete_area_files($context->id, 'tool_courserating', 'review');
+        }
         $DB->execute('DELETE from {' . flag::TABLE . '} WHERE ratingid IN (SELECT id FROM {' .
             rating::TABLE . '} WHERE courseid = ?)', [$courseid]);
         $DB->delete_records(rating::TABLE, ['courseid' => $courseid]);