+++ tests/privacy/provider_test.php
@@ -71,7 +71,7 @@ final class provider_test extends \core_privacy\tests\provider_testcase {
      */
     public function test_export_user_data(): void {
 
-        [$user, $course] = $this->setup_test_scenario_data();
+        [$user, $course, , , $rating2] = $this->setup_test_scenario_data();
         $coursectx = \context_course::instance($course->id);
         $this->setAdminUser();
 
@@ -97,24 +97,81 @@ final class provider_test extends \core_privacy\tests\provider_testcase {
         // Test the tool_cohortroles data is not exported at the course context level.
         $writer = writer::with_context($coursectx);
         $this->assertTrue($writer->has_any_data());
-        $this->assertNotEmpty($writer->get_data(['Course ratings', $course->shortname]));
+        $data = $writer->get_data(['Course ratings', $course->shortname]);
+        $this->assertNotEmpty($data);
+        $this->assertEquals(4, $data->rating);
+
+        // The review files are exported and the links to them are rewritten.
+        $this->assertStringNotContainsString('@@PLUGINFILE@@', $data->review);
+        $this->assertStringContainsString('files/image1.png', $data->review);
+        $files = $writer->get_files(['Course ratings', $course->shortname]);
+        $this->assertEquals(['image1.png'], array_keys($files));
+
+        // The flags that the user placed on other users' reviews are exported.
+        $data = $writer->get_data(['Course ratings', $course->shortname, 'Flagged reviews']);
+        $this->assertCount(1, $data->flags);
+        $this->assertEquals($rating2->get('id'), $data->flags[0]->ratingid);
     }
 
     /**
      * Set up scenario data
      *
-     * @return array
+     * Two users rate the same course, both reviews have embedded files and each user flags the review of the other user.
+     *
+     * @return array [$user1, $course, $user2, $rating1, $rating2]
      */
     protected function setup_test_scenario_data() {
         $user1 = $this->getDataGenerator()->create_user();
         $user2 = $this->getDataGenerator()->create_user();
         $course = $this->getDataGenerator()->create_course(['shortname' => 'c1']);
         $this->setUser($user2);
-        $r = \tool_courserating\api::set_rating($course->id, (object)['rating' => 5]);
+        $rating2 = \tool_courserating\api::set_rating($course->id, (object)['rating' => 5]);
+        $this->add_review_file($rating2, 'image2.png');
         $this->setUser($user1);
-        \tool_courserating\api::set_rating($course->id, (object)['rating' => 4]);
-        api::flag_review($r->get('id'));
-        return [$user1, $course];
+        $rating1 = \tool_courserating\api::set_rating($course->id, (object)['rating' => 4]);
+        $this->add_review_file($rating1, 'image1.png');
+        api::flag_review($rating2->get('id'));
+        $this->setUser($user2);
+        api::flag_review($rating1->get('id'));
+        return [$user1, $course, $user2, $rating1, $rating2];
+    }
+
+    /**
+     * Add a file to the review and embed it in the review text
+     *
+     * @param rating $rating
+     * @param string $filename
+     */
+    protected function add_review_file(rating $rating, string $filename): void {
+        $context = \context_course::instance($rating->get('courseid'));
+        get_file_storage()->create_file_from_string([
+            'contextid' => $context->id,
+            'component' => 'tool_courserating',
+            'filearea' => 'review',
+            'itemid' => $rating->get('id'),
+            'filepath' => '/',
+            'filename' => $filename,
+        ], 'Image content');
+        $rating->set('review', '<p>Review <img src="@@PLUGINFILE@@/' . $filename . '" alt="Image"></p>');
+        $rating->save();
+    }
+
+    /**
+     * Get the names of the files in the review file area of the course
+     *
+     * @param \stdClass $course
+     * @return array
+     */
+    protected function get_review_files(\stdClass $course): array {
+        $files = get_file_storage()->get_area_files(
+            \context_course::instance($course->id)->id,
+            'tool_courserating',
+            'review',
+            false,
+            'filename',
+            false
+        );
+        return array_values(array_map(fn($f) => $f->get_filename(), $files));
     }
 
     /**
@@ -127,9 +184,12 @@ final class provider_test extends \core_privacy\tests\provider_testcase {
         $coursectx = \context_course::instance($course->id);
         $this->setAdminUser();
 
+        $this->assertEquals(['image1.png', 'image2.png'], $this->get_review_files($course));
+
         provider::delete_data_for_all_users_in_context($coursectx);
         $this->assertEmpty($DB->get_records(rating::TABLE));
         $this->assertEmpty($DB->get_records(flag::TABLE));
+        $this->assertEmpty($this->get_review_files($course));
     }
 
     /**
@@ -138,17 +198,39 @@ final class provider_test extends \core_privacy\tests\provider_testcase {
     public function test_delete_data_for_user(): void {
         global $DB;
 
-        [$user, $course] = $this->setup_test_scenario_data();
+        [$user, $course, $user2, $rating1, $rating2] = $this->setup_test_scenario_data();
         $coursectx = \context_course::instance($course->id);
+
+        // The user also has a rating in another course, it should not be deleted.
+        $course2 = $this->getDataGenerator()->create_course();
+        $this->setUser($user);
+        $rating3 = api::set_rating($course2->id, (object)['rating' => 3]);
+        $this->add_review_file($rating3, 'image3.png');
         $this->setAdminUser();
 
         // Test the User's retrieved contextlist contains two contexts.
         $contextlist = provider::get_contexts_for_userid($user->id);
         $contexts = $contextlist->get_contexts();
-        $this->assertCount(1, $contexts);
+        $this->assertCount(2, $contexts);
 
         $approvedcontextlist = new approved_contextlist($user, 'tool_courserating', [$coursectx->id]);
         provider::delete_data_for_user($approvedcontextlist);
+
+        // The user's rating, the files in their review and all flags on it are deleted,
+        // as well as the user's flags on the other reviews.
+        $this->assertEqualsCanonicalizing(
+            [$rating2->get('id'), $rating3->get('id')],
+            $DB->get_fieldset_select(rating::TABLE, 'id', '1=1')
+        );
+        $this->assertEmpty($DB->get_records(flag::TABLE));
+        $this->assertEquals(['image2.png'], $this->get_review_files($course));
+        $this->assertEquals(['image3.png'], $this->get_review_files($course2));
+
+        // The other user's data remains.
+        $contextlist = provider::get_contexts_for_userid($user2->id);
+        $this->assertEqualsCanonicalizing([$coursectx->id], $contextlist->get_contextids());
+        $contextlist = provider::get_contexts_for_userid($user->id);
+        $this->assertEqualsCanonicalizing([\context_course::instance($course2->id)->id], $contextlist->get_contextids());
     }
 
     /**
@@ -189,5 +271,28 @@ final class provider_test extends \core_privacy\tests\provider_testcase {
         $userlist1 = new \core_privacy\local\request\userlist($coursectx, $component);
         provider::get_users_in_context($userlist1);
         $this->assertCount(0, $userlist1);
+        $this->assertEmpty($this->get_review_files($course));
+    }
+
+    /**
+     * Test that data for one user in approved userlist is deleted together with the flags on their rating.
+     */
+    public function test_delete_data_for_users_one_user(): void {
+        global $DB;
+        $component = 'tool_courserating';
+
+        [$user, $course, $user2, $rating1, $rating2] = $this->setup_test_scenario_data();
+        $coursectx = \context_course::instance($course->id);
+        $this->setAdminUser();
+
+        $approvedlist = new approved_userlist($coursectx, $component, [$user->id]);
+        provider::delete_data_for_users($approvedlist);
+
+        $userlist = new \core_privacy\local\request\userlist($coursectx, $component);
+        provider::get_users_in_context($userlist);
+        $this->assertEqualsCanonicalizing([$user2->id], $userlist->get_userids());
+        $this->assertEquals([$rating2->get('id')], $DB->get_fieldset_select(rating::TABLE, 'id', '1=1'));
+        $this->assertEmpty($DB->get_records(flag::TABLE));
+        $this->assertEquals(['image2.png'], $this->get_review_files($course));
     }
 }